+++ e2e-tests/tests/sveltekit-hooks.test.ts
@@ -241,6 +241,10 @@ describe('Sveltekit', () => {
 
   // If you don't want to use Session Replay, just remove the line below:
   integrations: [replayIntegration()],
+
+  // Enable sending user PII (Personally Identifiable Information)
+  // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+  sendDefaultPii: true,
 });`,
         'export const handleError = handleErrorWithSentry(',
       ]);
@@ -256,6 +260,11 @@ describe('Sveltekit', () => {
 
   // Enable logs to be sent to Sentry
   enableLogs: true,
+
+
+  // Enable sending user PII (Personally Identifiable Information)
+  // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+  sendDefaultPii: true,
 
   // uncomment the line below to enable Spotlight (https://spotlightjs.com)
   // spotlight: import.meta.env.DEV,
@@ -308,7 +317,8 @@ describe('Sveltekit', () => {
     replaysSessionSampleRate: 0.1,
     replaysOnErrorSampleRate: 1,
     integrations: [Sentry.replayIntegration()],
-    enableLogs: true
+    enableLogs: true,
+    sendDefaultPii: true
 })`,
         'export const handleError = Sentry.handleErrorWithSentry(',
       ]);
@@ -320,7 +330,8 @@ describe('Sveltekit', () => {
         `Sentry.init({
     dsn: "${TEST_ARGS.PROJECT_DSN}",
     tracesSampleRate: 1,
-    enableLogs: true
+    enableLogs: true,
+    sendDefaultPii: true
 })`,
         'export const handleError = Sentry.handleErrorWithSentry();',
       ]);