+++ src/sveltekit/templates.ts
@@ -39,6 +39,10 @@ ${
   integrations: [replayIntegration()],`
     : ''
 }
+
+  // Enable sending user PII (Personally Identifiable Information)
+  // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+  sendDefaultPii: true,
 });
 
 // If you have a custom error handler, pass it to \`handleErrorWithSentry\`
@@ -74,6 +78,11 @@ ${
 `
     : ''
 }
+
+  // Enable sending user PII (Personally Identifiable Information)
+  // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+  sendDefaultPii: true,
+
   // uncomment the line below to enable Spotlight (https://spotlightjs.com)
   // spotlight: import.meta.env.DEV,
 });`